backend/companion_app/requirements.txt
numpy>=1.26.0,<3.0.0
torch==2.1.0
huggingface-hub==0.16.4
sentence-transformers==2.2.2
fastapi>=0.100.0,<0.110.0
uvicorn>=0.22.0,<0.23.0
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
duckduckgo-search==3.8.0
litellm>=1.53.1
pydantic>=2.10,<3.0.0
sqlalchemy>=2.0.23,<3.0.0
psycopg2-binary==2.9.9
starlette>=0.27.0,<0.28.0
httpx>=0.24.1,<0.25.0
typing-extensions>=4.7.1,<5.0.0
crawl4ai==0.4.245
